Core/Opcode: more clean-up and work done on this system.
this also fixes build.

Note: we need more hands on opcodes and get these missing ones named and inplace... Reversers and sniffers rollcall guys... let's finish this 406 where it was left off for this part of the system.

+++ src/server/game/Server/Protocol/Handlers/SpellHandler.cpp
@@ -637,7 +637,7 @@ void WorldSession::HandleMirrorImageDataRequest(WorldPacket & recv_data)
     else
     {
         // Skip player data for creatures
-        data << uint8(0);
+        data << uint32(0);
         data << uint32(0);
         data << uint32(0);
         data << uint32(0);
